development/requirements.txt
backports.weakref==1.0rc1
bcolz==1.1.2
bleach==1.5.0
cycler==0.10.0
decorator==4.0.11
h5py==2.7.0
html5lib==0.9999999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.5
Markdown==2.6.8
matplotlib==2.0.2
numpy==1.13.0
olefile==0.44
pandas==0.20.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.0
prompt-toolkit==1.0.14
protobuf==3.3.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
scikit-learn==0.18.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
tensorflow-gpu==1.2.1
Theano==0.9.0
tornado==4.5.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
